js update

+++ db/migrate/20111208033458_add_geolocation_to_building.rb
@@ -1,7 +1,7 @@
 class AddGeolocationToBuilding < ActiveRecord::Migration
   def self.up
-    add_column :buildings, :lat, :double
-    add_column :buildings, :lon, :double
+    add_column :buildings, :lat, :decimal
+    add_column :buildings, :lon, :decimal
   end
 
   def self.down
